+++ html/src/com/angian/libgdx/pipes/client/HtmlLauncher.java
@@ -1,0 +1,22 @@
+package com.angian.libgdx.pipes.client;
+
+import com.badlogic.gdx.ApplicationListener;
+import com.badlogic.gdx.backends.gwt.GwtApplication;
+import com.badlogic.gdx.backends.gwt.GwtApplicationConfiguration;
+import com.angian.libgdx.pipes.PipesGame;
+
+public class HtmlLauncher extends GwtApplication {
+
+        @Override
+        public GwtApplicationConfiguration getConfig () {
+                // Resizable application, uses available space in browser
+                //return new GwtApplicationConfiguration(true);
+                // Fixed size application:
+                return new GwtApplicationConfiguration(1200, 900);
+        }
+
+        @Override
+        public ApplicationListener createApplicationListener () {
+                return new PipesGame();
+        }
+}